chore(tickets): change skeleton components dir
